# DOC: Use vanilla MathJax distribution, fall back to vendored `scipy-mathjax` if not available