<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue

This change was proposed in the SciPy Community Call (Americas) on 17/04/2024 (please read through the [meeting notes](https://hackmd.io/pyhudrC5TgSwdJtpPldHgQ#Agenda-for-April-17th-2024) for more context about this change).

#### What does this implement/fix?

This PR adds a small code snippet to check for an internet connection by pinging Google's DNS at `8.8.8.8`. If an internet connection is available, it will use the vanilla MathJax distribution, and if there isn't, then the local (vendored) MathJax distribution will be used. This ensures that the local one is used as just a fallback with a mechanism to check for the
availability of the usual distribution.

#### Additional information

Here are the three possible scenarios:

1. Building the documentation for deployment via GitHub Pages or on CircleCI: the vanilla MathJax distribution will be used
2. Building the documentation locally with an internet connection: the vanilla MathJax distribution will be used
3. Building the documentation locally without an internet connection: the vendored MathJax distribution will be used

The third scenario is quite rare and will affect just contributors and developers for SciPy (and those without access to the internet). Therefore, the vendored MathJax (`scipy-mathjax`) will be used only without the presence of an internet connection – in the case of the first scenario, GitHub Actions runners used to deploy to GitHub Pages and runners on CircleCI both have access to the internet.

To not produce any breaking changes and keep behaviour consistent with what SciPy has been doing in the past (starting with gh-7376), I think it would be great to continue using the vendored distribution for the official SciPy docs, so if there is a handy environment variable that can be used to check if they are being deployed as a part of a release, I shall modify the logic here to include that (and therefore refine scenario 1).

xref: https://github.com/executablebooks/MyST-NB/issues/590 for why this change is being proposed – it would be better to resolve this issue, so this PR is being opened just for visibility.

cc: @melissawm