@agriyakhetarpal 2024-04-18
The sdist + wheel build on Windows (without Pythran acceleration) looks like it was a [GitHub runner error](https://github.com/scipy/scipy/actions/runs/8740069512) and should be unrelated