@DietBru 2024-04-21
Please note that the default MathML renderer (CHTML) has an annoying bug: Depending on the zoom level the over-line markings are suppressed.
On my computer (Firefox in KDE Wayland on 1920x1080 display) it happens at 100% zoom state. 
From the [Short-Time Fourier Transform Section](https://output.circle-artifacts.com/output/job/d9a7974e-fa91-4632-95d3-eadbcecc1112/artifacts/0/html/tutorial/signal.html#short-time-fourier-transform) with CHTML renderer:
![image](https://github.com/scipy/scipy/assets/12721170/c8dc5a48-9f2e-4828-a641-db90ace79b36)

With SVG renderer:
![image](https://github.com/scipy/scipy/assets/12721170/faebcbd5-fe08-4559-aefc-fc67c993f0a9)

The renderer can be changed by right clicking on the equation and opening the sub-menu `Math Settings` → `Math Renderer`.

In stable SciPy, the default is seems to be the SVG renderer.