@agriyakhetarpal 2024-04-22
Thanks for the comment, @DietBru – unfortunately on my macOS machine (resolution: 3024x1964 and the renderer being most likely Metal or XQuartz, not sure about them) and with the Chromium-based Brave Browser, full specification "Version 1.64.116 Chromium: 123.0.6312.105 (Official Build) (arm64)", I can't seem to reproduce that. I am able to view the over-line markings in the STFT section with both CHTML and SVG renderers:

<details>
<br>
<summary>CHTML</summary>

![STFT mathematical representation with CHTML renderer](https://github.com/scipy/scipy/assets/74401230/0d454514-5cf6-4e24-8b00-02eb6c7a1c41)

</details>

<details>
<br>
<summary>SVG</summary>

![STFT mathematical representation with SVG renderer](https://github.com/scipy/scipy/assets/74401230/da732ffa-25e6-44a1-be80-3ac07e0b7eb3)

</details>

The only difference that I can notice is that the SVG renderer is a bit less bold, but also slightly smaller, the latter may not be evident in the screenshots I posted because they are quite similar in dimensions (though I don't think that's relevant to what you mentioned).