@agriyakhetarpal 2024-04-22
> Thank you for the PR.
> 
> Do we know what is the impact on loading pages with a lot of maths (like in stats)? Could you do some timing? (With eg a trace from the dev tools of a browser.)
> 
> I am also concerned about the security implications of doing that. (We did have some CVE report about mathjax in the past.) Did other large projects have such discussions?

@tupui, I had looked at the impact on loading pages but couldn't find a truly reliable estimate for this last week through the in-browser developer tools (see below), so I'll additionally try to profile this with other tools available online (perhaps, something like http://www.webpagetest.org/). I will be looking into past CVE reports and the history of use of MathJax across SciPy and other projects.

<hr>

In the [Sphinx logs](https://app.circleci.com/pipelines/github/scipy/scipy/26613/workflows/249a75f0-9bfa-4082-9ead-68137680e65f/jobs/89657?invite=true#step-104-293_30) for the [docs built](https://output.circle-artifacts.com/output/job/85dc2cbc-0362-4c38-8f64-5b1ad8be6210/artifacts/0/html/tutorial/stats/discrete.html) on the latest commit, the proposed change is working and we have the MathJax distribution being loaded from the CDN. I purged my browser's caches because I found that it was loading the MathJax fonts from disk, which wasn't useful, and that revealed that they get loaded between as low as 9ms to 30ms (most of them were on the higher side, though). On subsequent reloads of the same page or other pages, they are placed into either the memory or disk caches, and thus take between 0 to 2 ms to load.

To say more, I chose [the Discrete Statistical Distributions page from `scipy.stats`](https://scipy.github.io/devdocs/tutorial/stats/discrete.html) in the developer documentation as a benchmark, and the local MathJax distribution indeed loads everything much faster, in 0–2 ms, so this says more about how efficiently these resources are getting cached on further accesses of the webpage and it is only the first time that one can experience a short delay, which should be more prominent on throttled devices.

I tested on a "Slow 3G" network, which took 10 seconds for the script to load, while the devdocs took 3.5 seconds, which is surely going to be noticeable. On a "Fast 3G" network with caches disabled, the CDN takes ~2.5 seconds, and the local MathJax takes just about a second (on some occasions a twentieth of a second less), which is more bearable. I think this is enough information for us to determine the impact of loading from the CDN, so I would be happy to implement something for this point as mentioned above and will be open to suggestions:

>  I think it would be great to continue using the vendored distribution for the official SciPy docs, so if there is a handy environment variable that can be used to check if they are being deployed as a part of a release, I shall modify the logic here to include that (and therefore refine scenario 1).

because the official docs are receiving the maximum number of views; on devdocs or CircleCI, we can use the MathJax CDN? I have yet to investigate plus verify the impacts of CVEs and use cases of the CDN by other projects, plus, this is only a temporary solution for the aforementioned MyST-NB issue anyway (which would be good to get fixed for other projects too), so the experiment has most likely been useful to understand how varying levels of network speeds can affect load times for scripts, regardless of whether this PR ends up being merged or not. It can be reiterated that this is for the first load and that these high load times do not show up for any type of throttled connection upon reloads, so the drawbacks highlighted may turn up to be not major.