@tupui 2024-04-22
Thank you for the analysis 👏 

It could be fine as we mostly serve fast connections. Though adding more latency could make us lose more ground.

@melissawm what do you think? From the notes, I think this was discussed with you in the call.