@agriyakhetarpal 2024-05-01
It is to be noted that the custom MathJax problem, which was the root of everything as noted in #20303, is fixed[^1] by adding the following line in `conf.py`:

```python
myst_update_mathjax = False
```

This configuration value can be evaluated to see if it breaks something else (I don't think it does), but this also reveals the source of the upstream bug. While this PR may not serve much purpose afterwards, I would still recommend updating the custom MathJax distribution to v3 for reasons mentioned above.

cc: @melissawm

[^1]: This is as noted in https://myst-parser.readthedocs.io/en/latest/syntax/optional.html#mathjax-and-math-parsing