@agriyakhetarpal 2024-05-06
Thank you for summarising, @melissawm!

> * If we use `myst_update_mathjax = False` in our conf.py, the immediate "mathjax not rendering in html pages originating from notebooks" is solved now;

Yes ✅

> * It is recommended that we update our custom mathjax package to use MathJax v3 at some point soon;

Indeed – I shall open another PR for doing this.

> * The full solution of this problem, involving a PR to upstream myst-nb, can be done at a later step.
>   Is that right? If so, do we still need this PR?

Yes. I shall add a comment to the existing issue on MyST-NB, notifying others, in case it has been triaged, that the bug has been isolated to a specific feature there. We do not need this PR anymore. I shall start working on updating MathJax to v3 – in the meantime, I will close this PR and link to these discussions in the newer PR.